+++ tests/unit/client/feedback/dataset/remote/test_dataset.py
@@ -19,7 +19,7 @@ from uuid import uuid4
 
 import httpx
 import pytest
-from argilla import FeedbackDataset, Workspace
+from argilla import FeedbackDataset, FeedbackRecord, Workspace
 from argilla.client.feedback.dataset.remote.dataset import RemoteFeedbackDataset
 from argilla.client.feedback.schemas import SuggestionSchema
 from argilla.client.feedback.schemas.remote.fields import RemoteTextField
@@ -137,7 +137,7 @@ class TestSuiteRemoteDataset:
 
         mock_httpx_client.patch.assert_called_once_with(
             url=f"/api/v1/datasets/{test_remote_dataset.id}/records",
-            json={"items": [{"id": str(test_remote_record.id), "metadata": {"new": "metadata"}, "suggestions": []}]},
+            json={"items": [{"id": str(test_remote_record.id), "suggestions": [], "metadata": {"new": "metadata"}}]},
         )
 
     def test_update_multiple_records(
@@ -273,4 +273,37 @@ class TestSuiteRemoteDataset:
         mock_httpx_client.post.assert_called_once_with(
             url=f"/api/v1/datasets/{test_remote_dataset.id}/vectors-settings",
             json={"name": expected_name, "title": expected_name, "dimensions": expected_dimensions},
+        )
+
+    def test_add_records_with_vectors(
+        self,
+        mock_httpx_client: httpx.Client,
+        test_remote_dataset: RemoteFeedbackDataset,
+    ) -> None:
+        mock_routes = {
+            "post": {f"/api/v1/datasets/{test_remote_dataset.id}/records": httpx.Response(status_code=204)},
+            "get": {
+                "/api/me": httpx.Response(
+                    status_code=200,
+                    content=UserModel(
+                        id=uuid4(),
+                        first_name="test",
+                        username="test",
+                        role=UserRole.owner,
+                        api_key="api.key",
+                        inserted_at=datetime.utcnow(),
+                        updated_at=datetime.utcnow(),
+                    ).json(),
+                ),
+                f"/api/v1/me/datasets/{test_remote_dataset.id}/metadata-properties": httpx.Response(
+                    status_code=200, json={"items": []}
+                ),
+            },
+        }
+        configure_mock_routes(mock_httpx_client, mock_routes)
+        test_remote_dataset.add_records(FeedbackRecord(fields={"text": "test"}, vectors={"test": [1, 2, 3]}))
+
+        mock_httpx_client.post.assert_called_once_with(
+            url=f"/api/v1/datasets/{test_remote_dataset.id}/records",
+            json={"items": [{"fields": {"text": "test"}, "suggestions": [], "vectors": {"test": [1, 2, 3]}}]},
         )